Approximates the hybrid xsd to the v003

+++ api-generator/generator-java/src/main/java/esa/mo/tools/stubgen/GeneratorJava.java
@@ -281,7 +281,7 @@ public class GeneratorJava extends GeneratorLangs {
 
     @Override
     public void createListClass(File folder, AreaType area, ServiceType service,
-            String srcTypeName, boolean isAbstract, Long shortFormPart) throws IOException {
+            String srcTypeName, boolean isAbstract, Integer shortFormPart) throws IOException {
         JavaLists javaLists = new JavaLists(this);
         String listName = srcTypeName + "List";
